Simplified
Basically I've removed all the eye candy. There's only one image, which
is served from gravatar and everything is inline execpt the that
gravatar image and my PGP key. I could've used data url to make the PGP
key inline but that was not done becase the gist provides a managible
and updateable way of sharing PGP key.